pyproject.toml
[project]
name = "priorflow"
version = "0.1.0"
description = "AI-powered prior authorization agent using Browser Use"
requires-python = ">=3.12"
dependencies = [
    "browser-use>=0.1.40",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.setuptools.packages.find]
include = ["shared*", "agents*", "tools*", "server*"]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
